requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
argparse; python_version>="3.0"
pygame>=2.5.0

